requirements.txt
asgiref==3.6.0
attrs==22.2.0
coverage==7.1.0
Django==4.1.6
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
Faker==17.0.0
iniconfig==2.0.0
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
